Validate length of multibyte characters better
If we length check on just the string when there's some foreign characters in there it can often come back as a length you wouldn't expect. This doesn't make much sense to me for validation purposes, you always want the length of a string to mean x characters, not bytes.

See my Go Playground: https://play.golang.org/p/GNDIG2Tzx4y

I would expect the behaviour of the runed string, not the latter when validating.